requirements.txt
numpy
moderngl
pygame-ce